+++ wordcount-java/run_tracing.sh
@@ -1,7 +1,17 @@
 #!/bin/bash
 SCRIPTDIR="$( cd "$( dirname "${BASH_SOURCE[0]}" )" >/dev/null && pwd )"
+grn=$'\e[1;32m'
+blu=$'\e[1;34m'
+red=$'\e[1;91m'
+end=$'\e[0m'
+function printError { 
+  echo "${red}======== $1 ========${end}"
+}
+function printMsg { 
+  echo "${grn}======== $1 ========${end}"
+}
 
-### START DATACLAY WITH TRACING ####
+printMsg "Starting dataClay with tracing"
 export COMMAND_OPTS="--tracing"
 pushd $SCRIPTDIR/dataclay
 docker-compose kill
@@ -11,18 +21,19 @@ popd
 
 pushd $SCRIPTDIR
 
-### BUILD ####
+printMsg "Building app"
 docker build --network=dataclay_default \
 	--build-arg CACHEBUST=$(date +%s) \
 	-t bscdataclay/wordcount-java-demo .			
+if [ $? -ne 0 ]; then printError "DEMO FAILED"; exit 1; fi
 	
-### RUN DEMO ####
-
+printMsg "Running demo"
 mkdir -p trace
 # Generate
 docker run --network=dataclay_default \
 	-v `pwd`/app/text:/usr/src/demo/app/text:ro \
     bscdataclay/wordcount-java-demo -Dexec.mainClass="TextCollectionGen" words /usr/src/demo/app/text
+if [ $? -ne 0 ]; then printError "DEMO FAILED"; exit 1; fi
 
 # Word count
 # Modify session configuration to add flag Tracing=True and mount trace volume to collect traces once done
@@ -31,10 +42,14 @@ docker run --network=dataclay_default \
 	-v `pwd`/app/text:/usr/src/demo/app/text:ro \
     -v `pwd`/trace:/usr/src/demo/app/trace:rw \
     bscdataclay/wordcount-java-demo --tracing -Dexec.mainClass="Wordcount" words
+if [ $? -ne 0 ]; then printError "DEMO FAILED"; exit 1; fi
 
-echo "Traces created at $(pwd)/trace/"
+printMsg "Traces created at $(pwd)/trace/"
 
-### STOP DATACLAY ####
+printMsg "Stopping dataClay"
 pushd $SCRIPTDIR/dataclay
 docker-compose down -v
 popd
+
+echo ""
+printMsg " DEMO SUCCESSFULLY FINISHED :) "
