+++ wordcount-python/run_demo.sh
@@ -1,7 +1,16 @@
 #!/bin/bash
 SCRIPTDIR="$( cd "$( dirname "${BASH_SOURCE[0]}" )" >/dev/null && pwd )"
-
-### START DATACLAY ####
+grn=$'\e[1;32m'
+blu=$'\e[1;34m'
+red=$'\e[1;91m'
+end=$'\e[0m'
+function printError { 
+  echo "${red}======== $1 ========${end}"
+}
+function printMsg { 
+  echo "${grn}======== $1 ========${end}"
+}
+printMsg "Starting dataClay"
 pushd $SCRIPTDIR/dataclay
 docker-compose kill
 docker-compose down -v #sanity check
@@ -9,18 +18,23 @@ docker-compose up -d
 popd
 
 pushd $SCRIPTDIR
-### BUILD ####
+printMsg "Building app"
 docker build --network=dataclay_default \
 	--build-arg CACHEBUST=$(date +%s) \
 	-t bscdataclay/wordcount-python-demo .			
+if [ $? -ne 0 ]; then printError "DEMO FAILED"; exit 1; fi
 
-### RUN DEMO ####
+printMsg "Running demo"
 docker run --network=dataclay_default \
     bscdataclay/wordcount-python-demo src/wordcount.py
+if [ $? -ne 0 ]; then printError "DEMO FAILED"; exit 1; fi
     
 popd
 
-### STOP DATACLAY ####
+printMsg "Stopping dataClay"
 pushd $SCRIPTDIR/dataclay
 docker-compose down -v
 popd
+
+echo ""
+printMsg " DEMO SUCCESSFULLY FINISHED :) "
